requirements.txt

imutils
keras
matplotlib
numpy
opencv-python
scikit-image
scikit-learn

# tensorflow
tensorflow-macos
tensorflow-metal
